backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
numpy
Pillow
anthropic
shapely
